refactor: moved mocking code to a module
