+++ mock/mockdata.go
@@ -1,11 +1,11 @@
-package services
+package mock
 
 import (
 	"encoding/json"
 	"os"
 )
 
-func getMockData[T any](mockPath string) (map[string]T, error) {
+func GetMockData[T any](mockPath string) (map[string]T, error) {
 	// TODO: save to mock data to db
 	f, err := os.ReadFile(mockPath)
 	if err != nil {
